+++ src/cpp/session/SessionClientEvent.cpp
@@ -209,7 +209,6 @@ const int kMemoryUsageChanged = 191;
 const int kCommandCallbacksChanged = 192;
 const int kConsoleActivate = 193;
 const int kJobsActivate = 194;
-const int kOpenSourceFile = 195;
 }
 
 void ClientEvent::init(int type, const json::Value& data)
@@ -583,8 +582,6 @@ std::string ClientEvent::typeName() const
          return "console_activate";
       case client_events::kJobsActivate:
          return "jobs_activate";
-      case client_events::kOpenSourceFile:
-         return "open_source_file";
       default:
          LOG_WARNING_MESSAGE("unexpected event type: " + 
                              safe_convert::numberToString(type_));