+++ src/cpp/session/include/session/SessionModuleContext.hpp
@@ -125,6 +125,9 @@ bool isPdfLatexInstalled();
 // is the file a text file
 bool isTextFile(const core::FilePath& targetPath);
 
+// edit a file
+void editFile(const core::FilePath& targetPath, int lineNumber = -1);
+
 // find the location of the R script
 core::Error rBinDir(core::FilePath* pRBinDirPath);
 core::Error rScriptPath(core::FilePath* pRScriptPath);
@@ -917,6 +920,7 @@ struct QuartoConfig
    std::string project_dir;
    std::string project_output_dir;
    std::vector<std::string> project_formats;
+   std::vector<std::string> project_bibliographies;
 };
 
 QuartoConfig quartoConfig(bool refresh = false);