+++ src/cpp/session/include/session/worker_safe/session/SessionClientEvent.hpp
@@ -210,7 +210,6 @@ extern const int kMemoryUsageChanged;
 extern const int kCommandCallbacksChanged;
 extern const int kConsoleActivate;
 extern const int kJobsActivate;
-extern const int kOpenSourceFile;
 }
    
 class ClientEvent