+++ src/cpp/session/modules/SessionSource.cpp
@@ -1295,7 +1295,10 @@ void enqueFileEditEvent(const std::string& file)
    fileJson["mime_type"] = filePath.getMimeContentType();
    
    // fire event
-   ClientEvent event(client_events::kFileEdit, fileJson);
+   json::Object eventJson;
+   eventJson["file"] = fileJson;
+   eventJson["position"] = json::Value();
+   ClientEvent event(client_events::kFileEdit, eventJson);
    module_context::enqueClientEvent(event);
 }